Move functions to .cpp file
